move the MCP3008 code out of the board package
