+++ components/board/mcp3008helper/mcp3008.go
@@ -1,4 +1,6 @@
-package board
+// Package mcp3008helper is shared code for hooking an MCP3008 ADC up to a board. It is used in
+// both the pi and genericlinux board implementations, but does not implement a board directly.
+package mcp3008helper
 
 import (
 	"context"
